requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
pymongo==4.3.3
python-dotenv==1.0.0
flask-paginate==2022.1.8
Flask-JWT-Extended==4.4.4
Flask-Script==2.0.6
Flask-MySQLdb==1.0.1
passlib==1.7.4